+++ client-cli.c
@@ -130,7 +130,7 @@ static void write_preference(const char *name, const char *value) {
 		if(len == -2) continue;
 		if(len == 0 || *buffer == '#') continue;
 		if(strncmp(buffer, name_equal, name_len + 1) == 0) {
-			if(len == name_len + 1 + value_len) {
+			if((size_t)len == name_len + 1 + value_len) {
 				int n = fseek(preference_file, -1, SEEK_CUR) == 0 && fgetc(preference_file) == '\n';
 				if(fseek(preference_file, -n - value_len, SEEK_CUR) < 0) return;
 				fputs(value, preference_file);
